Add all types, use plain objects instead of entities

+++ src/bot/flows/tagging.js
@@ -1,26 +1,33 @@
-import { Tag } from '../../tags/Tag.js'
+/** @typedef {import('telegraf').Context} Context */
 
+/**
+ * @param {{
+ *   userSessionRepository: import('../../users/DynamodbUserSessionRepository').DynamodbUserSessionRepository
+ *   queuedStickerRepository: import('../../queue/DynamodbQueuedStickerRepository').DynamodbQueuedStickerRepository
+ *   tagRepository: import('../../tags/DynamodbTagRepository').DynamodbTagRepository
+ *   sendNextQueuedSticker: Function
+ *   bot: import('telegraf').Telegraf
+ * }} input
+ */
 export function useTaggingFlow({ queuedStickerRepository, userSessionRepository, tagRepository, bot, sendNextQueuedSticker }) {
+  /** @param {Context} context */
   async function handleTag(context, next) {
+    if (!context.chat || !context.message) return
     if (context.message.text.startsWith('/')) return next()
 
     const { userId } = context.state
-    const { stickerSetName, stickerFileId, stickerFileUniqueId, stickerMessageId } = await userSessionRepository.getContext(userId)
+    const { sticker, stickerMessageId } = await userSessionRepository.getContext(userId)
 
-    if (!stickerSetName || !stickerFileId || !stickerFileUniqueId || !stickerMessageId) return
+    if (!sticker || !stickerMessageId) return
 
     const value = context.message.text.trim().toLowerCase()
     if (!value) return
 
-    await tagRepository.storeTag(
-      new Tag({
-        authorUserId: userId,
-        stickerFileUniqueId,
-        stickerFileId,
-        stickerSetName,
-        value,
-      })
-    )
+    await tagRepository.storeTag({
+      sticker,
+      authorUserId: userId,
+      value,
+    })
 
     await bot.telegram.editMessageReplyMarkup(context.chat.id, stickerMessageId, undefined, undefined)
     await context.reply(`✏️ The sticker has been tagged as "${value}"`)
@@ -28,37 +35,38 @@ export function useTaggingFlow({ queuedStickerRepository, userSessionRepository,
     await sendNextQueuedSticker(context)
   }
 
+  /** @param {Context} context */
   async function tagSingle(context) {
+    if (!context.chat) return
     context.deleteMessage().catch(() => {})
 
     const { userId } = context.state
-    const { stickerSetName, stickerFileUniqueId, stickerFileId, stickerMessageId } = await userSessionRepository.getContext(userId)
+    const { sticker, stickerMessageId } = await userSessionRepository.getContext(userId)
 
-    if (!stickerSetName || !stickerFileUniqueId || !stickerFileId || !stickerMessageId) return
+    if (!sticker || !stickerMessageId) return
 
     bot.telegram.deleteMessage(context.chat.id, stickerMessageId).catch(() => {})
 
     await enqueueStickers({
       context,
       userId,
-      stickerSetName,
-      stickers: [{
-        stickerFileId,
-        stickerFileUniqueId,
-      }],
+      stickers: [sticker],
     })
   }
 
+  /** @param {Context} context */
   async function tagUntagged(context) {
+    if (!context.chat) return
     context.deleteMessage().catch(() => {})
 
     const { userId } = context.state
-    const { stickerSetName, stickerMessageId } = await userSessionRepository.getContext(userId)
+    const { sticker, stickerMessageId } = await userSessionRepository.getContext(userId)
 
-    if (!stickerSetName || !stickerMessageId) return
+    if (!sticker || !stickerMessageId) return
 
     bot.telegram.deleteMessage(context.chat.id, stickerMessageId).catch(() => {})
 
+    const stickerSetName = sticker.setName
     const stickerSet = await bot.telegram.getStickerSet(stickerSetName)
     const stickerFileUniqueIds = stickerSet.stickers.map(sticker => sticker.file_unique_id)
     const statusMap = await tagRepository.queryTagStatus({ stickerFileUniqueIds })
@@ -66,26 +74,29 @@ export function useTaggingFlow({ queuedStickerRepository, userSessionRepository,
     await enqueueStickers({
       context,
       userId,
-      stickerSetName,
       stickers: stickerSet.stickers
         .map(sticker => ({
-          stickerFileId: sticker.file_id,
-          stickerFileUniqueId: sticker.file_unique_id
+          setName: stickerSetName,
+          fileId: sticker.file_id,
+          fileUniqueId: sticker.file_unique_id
         }))
-        .filter((sticker) => !statusMap[sticker.stickerFileUniqueId])
+        .filter((sticker) => !statusMap[sticker.fileUniqueId])
     })
   }
 
+  /** @param {Context} context */
   async function tagUntaggedByMe(context) {
+    if (!context.chat) return
     context.deleteMessage().catch(() => {})
 
     const { userId } = context.state
-    const { stickerSetName, stickerMessageId } = await userSessionRepository.getContext(userId)
+    const { sticker, stickerMessageId } = await userSessionRepository.getContext(userId)
 
-    if (!stickerSetName || !stickerMessageId) return
+    if (!sticker || !stickerMessageId) return
 
     bot.telegram.deleteMessage(context.chat.id, stickerMessageId).catch(() => {})
 
+    const stickerSetName = sticker.setName
     const stickerSet = await bot.telegram.getStickerSet(stickerSetName)
     const stickerFileUniqueIds = stickerSet.stickers.map(sticker => sticker.file_unique_id)
     const statusMap = await tagRepository.queryTagStatus({ stickerFileUniqueIds, authorUserId: userId })
@@ -93,24 +104,27 @@ export function useTaggingFlow({ queuedStickerRepository, userSessionRepository,
     await enqueueStickers({
       context,
       userId,
-      stickerSetName,
       stickers: stickerSet.stickers
         .map(sticker => ({
-          stickerFileId: sticker.file_id,
-          stickerFileUniqueId: sticker.file_unique_id
+          setName: stickerSetName,
+          fileId: sticker.file_id,
+          fileUniqueId: sticker.file_unique_id
         }))
-        .filter((sticker) => !statusMap[sticker.stickerFileUniqueId])
+        .filter((sticker) => !statusMap[sticker.fileUniqueId])
     })
   }
 
+  /** @param {Context} context */
   async function tagAll(context) {
+    if (!context.chat) return
     context.deleteMessage().catch(() => {})
 
     const { userId } = context.state
-    const { stickerSetName, stickerMessageId } = await userSessionRepository.getContext(userId)
+    const { sticker, stickerMessageId } = await userSessionRepository.getContext(userId)
 
-    if (!stickerSetName) return
+    if (!sticker || !stickerMessageId) return
 
+    const stickerSetName = sticker.setName
     const stickerSet = await bot.telegram.getStickerSet(stickerSetName)
 
     bot.telegram.deleteMessage(context.chat.id, stickerMessageId).catch(() => {})
@@ -118,23 +132,26 @@ export function useTaggingFlow({ queuedStickerRepository, userSessionRepository,
     await enqueueStickers({
       context,
       userId,
-      stickerSetName,
       stickers: stickerSet.stickers
         .map(sticker => ({
-          stickerFileId: sticker.file_id,
-          stickerFileUniqueId: sticker.file_unique_id
+          setName: stickerSetName,
+          fileId: sticker.file_id,
+          fileUniqueId: sticker.file_unique_id
         }))
     })
   }
 
-  async function enqueueStickers({ context, userId, stickerSetName, stickers }) {
+  /**
+   * @param {{
+   *   context: Context
+   *   userId: string
+   *   stickers: Sticker[]
+   * }} input 
+   */
+  async function enqueueStickers({ context, userId, stickers }) {
     await queuedStickerRepository.enqueue({
       userId,
-      stickers: stickers.map(({ stickerFileId, stickerFileUniqueId }) => ({
-        stickerSetName,
-        stickerFileId,
-        stickerFileUniqueId,
-      })),
+      stickers,
     })
 
     await sendNextQueuedSticker(context)
